+++ framework/model/src/main/java/is/codion/framework/model/package-info.java
@@ -22,8 +22,8 @@
  * <li>{@link is.codion.framework.model.EntityApplicationModel#USERNAME_PREFIX}
  * <li>{@link is.codion.framework.model.EntityApplicationModel#AUTHENTICATION_REQUIRED}
  * <li>{@link is.codion.framework.model.EntityApplicationModel#SAVE_DEFAULT_USERNAME}
- * <li>{@link is.codion.framework.model.EntityEditModel#PERSIST_FOREIGN_KEYS}
  * <li>{@link is.codion.framework.model.EntityEditModel#POST_EDIT_EVENTS}
+ * <li>{@link is.codion.framework.model.EntityEditModel.EntityEditor#PERSIST_FOREIGN_KEYS}
  * <li>{@link is.codion.framework.model.EntityTableModel#ON_INSERT}
  * <li>{@link is.codion.framework.model.EntityTableModel#HANDLE_EDIT_EVENTS}
  * <li>{@link is.codion.framework.model.ForeignKeyModelLink#SET_CONDITION_ON_INSERT}